// frontend/src/generated/models/Measurement.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tempmon API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Measurement
 */
export interface Measurement {
    /**
     * 
     * @type {number}
     * @memberof Measurement
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof Measurement
     */
    timestamp: string;
    /**
     * 
     * @type {number}
     * @memberof Measurement
     */
    temperature?: number | null;
    /**
     * 
     * @type {number}
     * @memberof Measurement
     */
    humidity?: number | null;
    /**
     * 
     * @type {string}
     * @memberof Measurement
     */
    ipAddress?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Measurement
     */
    macAddress?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Measurement
     */
    deviceName?: string | null;
}

/**
 * Check if a given object implements the Measurement interface.
 */
export function instanceOfMeasurement(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "timestamp" in value;

    return isInstance;
}

export function MeasurementFromJSON(json: any): Measurement {
    return MeasurementFromJSONTyped(json, false);
}

export function MeasurementFromJSONTyped(json: any, ignoreDiscriminator: boolean): Measurement {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'timestamp': json['timestamp'],
        'temperature': !exists(json, 'temperature') ? undefined : json['temperature'],
        'humidity': !exists(json, 'humidity') ? undefined : json['humidity'],
        'ipAddress': !exists(json, 'ipAddress') ? undefined : json['ipAddress'],
        'macAddress': !exists(json, 'macAddress') ? undefined : json['macAddress'],
        'deviceName': !exists(json, 'deviceName') ? undefined : json['deviceName'],
    };
}

export function MeasurementToJSON(value?: Measurement | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'timestamp': value.timestamp,
        'temperature': value.temperature,
        'humidity': value.humidity,
        'ipAddress': value.ipAddress,
        'macAddress': value.macAddress,
        'deviceName': value.deviceName,
    };
}